backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
pymysql==1.1.0
aiomysql==0.2.0
aiosqlite==0.19.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
authlib==1.3.0
cryptography==41.0.7

# Google OAuth
google-auth==2.26.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Validation & Serialization
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# AI Providers
openai==1.9.0
anthropic==0.12.0
google-generativeai==0.3.2
langchain==0.1.1
langchain-openai==0.0.3
langchain-community==0.0.10
tiktoken==0.5.2
faiss-cpu==1.7.4

# Caching & Queue
redis==5.0.1
aioredis==2.0.1
celery==5.3.6

# Rate Limiting
slowapi==0.1.9
fastapi-limiter==0.1.5

# HTTP & API
httpx==0.26.0
requests==2.31.0
aiohttp==3.9.3

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
python-dateutil==2.8.2
pytz==2023.3

# Storage & Files
boto3==1.34.13
pillow==10.2.0
python-magic==0.4.27

# Email
emails==0.6
aiosmtplib==3.0.1

# Monitoring & Logging
loguru==0.7.2
sentry-sdk[fastapi]==1.39.1
prometheus-client==0.19.0
python-logging-loki==0.3.1

# Testing
pytest==7.4.4
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.26.0
factory-boy==3.3.0
faker==22.0.0
pytest-mock==3.12.0

# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0
isort==5.13.2

# Production
gunicorn==21.2.0
supervisor==4.2.5
circus==0.18.0

# Database Migrations
yoyo-migrations==8.2.0